Migrate dployed to async/await
