+++ generate/src/contract/common.rs
@@ -1,6 +1,7 @@
 use crate::contract::Context;
 use crate::util::expand_doc;
-use proc_macro2::TokenStream;
+use ethcontract_common::Address;
+use proc_macro2::{Literal, TokenStream};
 use quote::quote;
 
 pub(crate) fn expand(cx: &Context) -> TokenStream {
@@ -14,6 +15,21 @@ pub(crate) fn expand(cx: &Context) -> TokenStream {
         .as_deref()
         .unwrap_or("Generated by `ethcontract`");
     let doc = expand_doc(doc_str);
+
+    let deployments = cx.deployments.iter().map(|(network_id, address)| {
+        let network_id = Literal::string(&network_id.to_string());
+        let address = expand_address(*address);
+
+        quote! {
+            artifact.networks.insert(
+                #network_id.to_owned(),
+                self::ethcontract::common::truffle::Network {
+                    address: #address,
+                    transaction_hash: None,
+                },
+            );
+        }
+    });
 
     quote! {
         #doc
@@ -31,8 +47,12 @@ pub(crate) fn expand(cx: &Context) -> TokenStream {
 
                 lazy_static! {
                     pub static ref ARTIFACT: Artifact = {
-                        Artifact::from_json(#artifact_json)
-                            .expect("valid artifact JSON")
+                        #[allow(unused_mut)]
+                        let mut artifact = Artifact::from_json(#artifact_json)
+                            .expect("valid artifact JSON");
+                        #( #deployments )*
+
+                        artifact
                     };
                 }
                 &ARTIFACT
@@ -86,8 +106,13 @@ pub(crate) fn expand(cx: &Context) -> TokenStream {
                 let web3 = Web3::new(transport);
                 let abi = Self::artifact().abi.clone();
                 let instance = Instance::with_transaction(web3, abi, address, transaction_hash);
-                let methods = Methods { instance };
 
+                Contract::from_raw(instance)
+            }
+
+            /// Creates a contract from a raw instance.
+            fn from_raw(instance: self::ethcontract::dyns::DynInstance) -> Self {
+                let methods = Methods { instance };
                 Contract { methods }
             }
 
@@ -134,5 +159,42 @@ pub(crate) fn expand(cx: &Context) -> TokenStream {
                     .finish()
             }
         }
+    }
+}
+
+/// Expands an `Address` into a literal representation that can be used with
+/// quasi-quoting for code generation.
+fn expand_address(address: Address) -> TokenStream {
+    let bytes = address
+        .as_bytes()
+        .iter()
+        .copied()
+        .map(Literal::u8_unsuffixed);
+
+    quote! {
+        self::ethcontract::H160([#( #bytes ),*])
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    #[rustfmt::skip]
+    fn expand_address_value() {
+        assert_quote!(
+            expand_address(Address::zero()),
+            {
+                self::ethcontract::H160([ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ])
+            },
+        );
+
+        assert_quote!(
+            expand_address("000102030405060708090a0b0c0d0e0f10111213".parse().unwrap()),
+            {
+                self::ethcontract::H160([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19])
+            },
+        );
     }
 }
