+++ src/contract.rs
@@ -11,8 +11,8 @@ use crate::errors::{DeployError, LinkError};
 use ethcontract_common::abi::{Error as AbiError, Result as AbiResult};
 use ethcontract_common::abiext::FunctionExt;
 use ethcontract_common::hash::H32;
-use ethcontract_common::truffle::Network;
 use ethcontract_common::{Abi, Artifact, Bytecode};
+use futures::compat::Future01CompatExt;
 use std::collections::HashMap;
 use std::hash::Hash;
 use web3::api::Web3;
@@ -94,7 +94,18 @@ impl<T: Transport> Instance<T> {
     /// Note that this does not verify that a contract with a matchin `Abi` is
     /// actually deployed at the given address.
     pub async fn deployed(web3: Web3<T>, artifact: Artifact) -> Result<Self, DeployError> {
-        DeployedFuture::new(web3, Deployments::new(artifact))
+        let network_id = web3.net().version().compat().await?;
+        let network = artifact
+            .networks
+            .get(&network_id)
+            .ok_or(DeployError::NotFound(network_id))?;
+
+        Ok(Instance::with_transaction(
+            web3,
+            artifact.abi,
+            network.address,
+            network.transaction_hash,
+        ))
     }
 
     /// Creates a contract builder with the specified `web3` provider and the
@@ -327,4 +338,71 @@ where
                 .map(move |(index, element)| (signature(element), (name.to_owned(), index)))
         })
         .collect()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::test::prelude::*;
+    use ethcontract_common::truffle::Network;
+    use ethcontract_common::Artifact;
+    use web3::types::H256;
+
+    #[test]
+    fn deployed() {
+        let mut transport = TestTransport::new();
+        let web3 = Web3::new(transport.clone());
+
+        let network_id = "42";
+        let address = addr!("0x0102030405060708091011121314151617181920");
+        let transaction_hash = Some(H256::repeat_byte(0x42));
+        let artifact = {
+            let mut artifact = Artifact::empty();
+            artifact.networks.insert(
+                network_id.to_string(),
+                Network {
+                    address,
+                    transaction_hash,
+                },
+            );
+            artifact
+        };
+
+        transport.add_response(json!(network_id)); // get network ID response
+        let instance = Instance::deployed(web3, artifact)
+            .immediate()
+            .expect("successful deployment");
+
+        transport.assert_request("net_version", &[]);
+        transport.assert_no_more_requests();
+
+        assert_eq!(instance.address(), address);
+        assert_eq!(instance.transaction_hash(), transaction_hash);
+    }
+
+    #[test]
+    fn deployed_not_found() {
+        let mut transport = TestTransport::new();
+        let web3 = Web3::new(transport.clone());
+
+        let network_id = "42";
+
+        transport.add_response(json!(network_id)); // get network ID response
+        let err = Instance::deployed(web3, Artifact::empty())
+            .immediate()
+            .expect_err("unexpected success getting deployed contract");
+
+        transport.assert_request("net_version", &[]);
+        transport.assert_no_more_requests();
+
+        assert!(
+            match &err {
+                DeployError::NotFound(id) => id == network_id,
+                _ => false,
+            },
+            "expected network {} not found error but got '{:?}'",
+            network_id,
+            err
+        );
+    }
 }
